requirements.txt
opencv-python
numpy
mxnet-cu100
scikit-learn==0.21.2
scikit-image
